feat(tests) Adding screenshot capture workflows
