+++ .github/workflows/pull-requests.yml
@@ -33,7 +33,7 @@ jobs:
       - name: Run the tests
         if: steps.cache-editor-tests.outputs.cache-hit != 'true'
         run: nix-shell --arg includeServerBuildSupport false --arg includeRunLocallySupport false --run check-editor-all-ci
-  
+
   deploy-staging:
     name: Deploy Staging Editor
     runs-on: ubuntu-latest
@@ -113,7 +113,7 @@ jobs:
           PERFORMANCE_GRAPHS_PLOTLY_USERNAME: ${{ secrets.PERFORMANCE_GRAPHS_PLOTLY_USERNAME}}
           PERFORMANCE_GRAPHS_PLOTLY_API_KEY: ${{secrets.PERFORMANCE_GRAPHS_PLOTLY_API_KEY}}
         run: |
-          nix-shell --arg includeServerBuildSupport false --arg includeRunLocallySupport false --run "cd performance-test; pnpm install --unsafe-perm; pnpm start"
+          nix-shell --arg includeServerBuildSupport false --arg includeRunLocallySupport false --run "cd puppeteer-tests; pnpm install --unsafe-perm; pnpm run performance-test"
       - name: Create comment
         uses: peter-evans/create-or-update-comment@v1
         if: ${{ success() || failure() }}
