+++ .github/workflows/screenshot-production.yml
@@ -1,0 +1,46 @@
+name: Screenshot Production
+on:
+  repository_dispatch:
+    types: [trigger-production-screenshot]
+  workflow_dispatch:
+  pull_request:
+
+jobs:
+  take-screenshot:
+    name: Take Screenshot
+    steps:
+      - name: Run Puppeteer
+        id: run-screenshot-test
+        env:
+          AWS_S3_BUCKET: ${{secrets.PERFORMANCE_GRAPHS_BUCKET}}
+          AWS_ACCESS_KEY_ID: ${{ secrets.PERFORMANCE_GRAPHS_ACCESS_KEY}}
+          AWS_SECRET_ACCESS_KEY: ${{ secrets.PERFORMANCE_GRAPHS_SECRET_KEY }}
+          AWS_REGION: ${{ secrets.STAGING_BUNDLE_REGION }}
+          BASE_EDITOR_URL: 'https://utopia.app'
+          PROJECT_ID: '9f8c4564'
+        run: |
+          nix-shell --arg includeServerBuildSupport false --arg includeRunLocallySupport false --run "cd puppeteer-tests; pnpm install --unsafe-perm; pnpm run screenshot-test"
+      - name: Build Discord Message
+        env:
+          TEMPLATE: >-
+            [
+              {
+                "title": "Production:",
+                "color": 14540253,
+                "image": {
+                  "url": $screenshot_url
+                }
+              },
+            ]
+          SCREENSHOT_URL: ${{ steps.run-screenshot-test.outputs.screenshot }}
+        run: |
+          echo "DISCORD_EMBEDS=$(jq -nc --arg screenshot_url "$SCREENSHOT_URL" "$TEMPLATE")" >> $GITHUB_ENV
+      - name: Send Discord Notification
+        uses: Ilshidur/action-discord@0.3.2
+        env:
+          DISCORD_WEBHOOK: ${{ secrets.DISCORD_BUILD_WEBHOOK }}
+          DISCORD_USERNAME: 'Puppeteer'
+          DISCORD_AVATAR: https://octodex.github.com/images/puppeteer.png
+          MESSAGE: 'Screenshot of latest Production deploy'
+        with:
+          args: ${{ env.MESSAGE }}
