+++ puppeteer-tests/.gitignore
@@ -1,3 +1,4 @@
 node_modules
 trace.json
 frameimages
+screenshot-*
