+++ puppeteer-tests/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "performance-test",
+  "name": "puppeteer-tests",
   "version": "1.0.0",
   "description": "",
-  "main": "src/index.ts",
   "scripts": {
-    "start": "ts-node src/index.ts",
+    "performance-test": "ts-node src/performance-test.ts",
+    "screenshot-test": "ts-node src/screenshot-test.ts",
     "preinstall": "npx only-allow pnpm"
   },
   "author": "",
@@ -17,13 +17,10 @@
     "multer": "1.4.2",
     "multer-s3": "2.9.0",
     "plotly": "1.0.6",
-    "puppeteer": "5.5.0",
+    "puppeteer": "10.4.0",
     "ts-node": "9.1.1",
     "typescript": "4.1.2",
     "uuid": "^8.3.2",
     "yn": "^4.0.0"
-  },
-  "devDependencies": {
-    "@types/puppeteer": "5.4.1"
   }
 }
